+++ components-ui/app/src/tol-ui/src/interfaces/index.ts
@@ -27,6 +27,7 @@ export * from "./Privilege";
 export * from "./Sunburst";
 export * from "./SocialViewer";
 export * from "./Table";
+export * from "./Tours";
 export * from "./TsDataSource";
 export * from "./TsDocParser";
 export * from "./User";